// ui/src/backend/generated/PlotConfig.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DatasetId } from "./DatasetId";
import type { PlotStyle } from "./PlotStyle";

/**
 * What is plotted. Columns are referenced by their unique display names.
 */
export type PlotConfig = { dataset_id: DatasetId, x: string, y: Array<string>, style: PlotStyle, };