// ui/src/backend/generated/ViewRequest.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DatasetId } from "./DatasetId";

export type ViewRequest = { dataset_id: DatasetId, x_min: number, x_max: number, y_min: number, y_max: number, width_px: number, height_px: number, };